cr, codegen, snapshot
