+++ apps/parsley/src/utils/logRow/index.ts
@@ -1,4 +1,4 @@
-import { RowType, SkippedLinesRow } from "types/logs";
+import { ProcessedLogLine, RowType, SkippedLinesRow } from "types/logs";
 /**
  *
  * @param start The start line number inclusive
@@ -10,4 +10,20 @@ const newSkippedLinesRow = (start: number, end: number): SkippedLinesRow => ({
   rowType: RowType.SkippedLines,
 });
 
-export { newSkippedLinesRow };
+/**
+ * `includesLineNumber` determines if a log line includes a given line number.
+ * @param logLine - the processed log line to check
+ * @param lineNumber - the line number to check
+ * @returns true if the log line includes the line number and false otherwise
+ */
+const includesLineNumber = (logLine: ProcessedLogLine, lineNumber?: number) => {
+  if (lineNumber === undefined) {
+    return false;
+  }
+  if (typeof logLine === "number") {
+    return logLine === lineNumber;
+  }
+  return lineNumber >= logLine.range.start && lineNumber < logLine.range.end;
+};
+
+export { includesLineNumber, newSkippedLinesRow };
