+++ apps/parsley/src/utils/logRowTypes/index.ts
@@ -27,25 +27,4 @@ const isSubsectionHeaderRow = (
 ): logLine is SubsectionHeaderRow =>
   typeof logLine === "object" && logLine.rowType === RowType.SubsectionHeader;
 
-/**
- * `includesLineNumber` determines if a log line includes a given line number.
- * @param logLine - the processed log line to check
- * @param lineNumber - the line number to check
- * @returns true if the log line includes the line number and false otherwise
- */
-const includesLineNumber = (logLine: ProcessedLogLine, lineNumber?: number) => {
-  if (lineNumber === undefined) {
-    return false;
-  }
-  if (typeof logLine === "number") {
-    return logLine === lineNumber;
-  }
-  return lineNumber >= logLine.range.start && lineNumber < logLine.range.end;
-};
-
-export {
-  isSectionHeaderRow,
-  isSkippedLinesRow,
-  isSubsectionHeaderRow,
-  includesLineNumber,
-};
+export { isSectionHeaderRow, isSkippedLinesRow, isSubsectionHeaderRow };
